Build core-only IRT blob
This builds a second IRT image called irt_core.nexe, which does not
include PPAPI.

This might be part of a future plan or a transitional stage, but the
related future choices are still undecided.  Regardless of how those
come out, this provides another option for experimentation.

BUG=none
TEST=trybots

R=bradchen@google.com,mseaborn@chromium.org,sehr@google.com

Review URL: http://codereview.chromium.org/6976005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5242 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2