ppapi sel_universal test migration: move geturl test from tests/ppapi_geturl/ tests/ppapi_simple_tests
This the last test that needed to be migrated.

The file.cc contains code for the two modes for downloading a file.
I separated those out into two classes to make the code more readable
at the expense of some code duplication.

The code for "streaming as file" is very verbose, so it can serve as an example.
Review URL: http://codereview.chromium.org/6995028

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5273 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2