New allocator for integrated runtime
This replaces the malloc in the integrated runtime with one that does not
rely on sbrk.  For this, we pull in the current public-domain dlmalloc
from the net.  It's verbatim from upstream except for a tiny addition of
conditionalization to avoid stdio dependencies, and some simple changes
to the hand-rolled spin-lock implementation (two previous CLs).

With this, the only use of sbrk left in the IRT is from
__pthread_initialize_minimal, which should be changed later to avoid it
entirely.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1784
TEST=trybots + no regressions in local {small,medium,large}_tests and chrome_browser_tests w/irt=1 and w/o

R=bradchen@google.com,mseaborn@chromium.org

Review URL: http://codereview.chromium.org/7016037

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5301 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2