Build IRT and its libraries independent of SDK library builds
Executive summary: we get an IRT image built under --nacl_glibc now.
Theoretically --nacl_glibc should no longer change the default to irt=0.
But the tests don't work right with glibc yet, so it's still left disabled.

This changes the scons rules to build the IRT nexe images and the libraries
that go into them separately from the main build.  They are always built
using the newlib toolchain, regardless of --nacl_glibc.  The libraries for
the IRT are built just for the IRT and go directly into its link, without
requiring the "partial SDK" installation step to find them.

The implicit exceptions are libnacl and libcrt_platform, which are not
proper dependencies in scons but brought in by default toolchain linking
rules.  Those will be taken from the newlib toolchain install unless scons
happens to build them first in the nacl_irt run.  We may try to clean that
up later in one way or another.

Currently this means several libraries can be built twice, once for the SDK
and once for the IRT.  Some of these libraries will eventually be
eliminated from the SDK and exist only in the IRT.

The only way we found to do this was to create a separate scons Environment
that wants to be a separate build mode.  To work around that and not
require every developer's and every bot's command lines to add "nacl_irt"
to their non-default --mode argument, we add a hack to add the new nacl_irt
mode whenever the nacl mode is being built and irt=1.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1805
TEST=trybots

R=bradchen@google.com,mseaborn@chromium.org,bradnelson@google.com

Review URL: http://codereview.chromium.org/7066054

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5546 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2