Dynamic loading: Move dyncode_delete() test to run on ARM
Move test_deleting_code() into dynamic_load_test.c so that it is run
on all architectures, including ARM.  When it was in
dynamic_modify_test.c it only ran on x86 because this file tests
features which are currently only implemented for x86.

This test is very weak, though.  It won't catch the issue that
dyncode_delete() does not flush the icache on ARM.  I will extend the
test in later changes.

Also remove test_deleting_code_disabled(), which is dead code.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=501
TEST=run_dynamic_load_test platform=arm

Review URL: http://codereview.chromium.org/7111001

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5558 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2