Make IRT -lppapi support user-provided main
This makes the ppapi_stub library (aka -lppapi) define PpapiPluginMain
and makes its main function an optional wrapper around that.  This
allows a nexe to define its own main and still work with the IRT.

The user_main test in ppapi_basic_object_browser_test now uses the IRT.

While we're at it, we clean up the IRT hook to start PPAPI to pass
back the return value of ppruntime's PpapiPluginMain so that return
value can be seen by the nexe calling the stub's PpapiPluginMain.
We also provide an alternate startup entrypoint PpapiPluginStart that
does not require the application to define magic symbol names used
with dangling references in the library.

BUG=cleanup
TEST=trybots

R=bradchen@google.com,nfullagar@google.com,mseaborn@chromium.org

Review URL: http://codereview.chromium.org/7104036

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5598 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2