Revamp runtime support for TLS
This revamps the internals of TLS support in the startup code and pthread
library.  The only material changes are to how we do things on ARM, but
everything is cleaned up a bit.

The way TLS support was implemented for ARM before was based on some past
misunderstandings of the ARM ABI for TLS and used some fragile kludges
because of the mismatch with what the compiler and linker expect the
runtime to be doing.  It tried to keep things more like the x86 version
than they are supposed to be.  The new implementation lays things out for
ARM the way it was intended, which matches the way that other ARM
implementation's like glibc's do it.  This allows us to get rid of the
kludge that adds a padding TLS variable in the startup code, for example.

The new support replaces the machine-specific parameterization functions
with a new set.  We intend these to be implemented as intrinsics in the
LLVM/pnacl compiler toolchain, but no such compiler support was actually
being used, so these changes do not conflict with the pnacl compiler.

The old machine-specific parameterization functions were declared in
nacl/tls.h alongside some machine-independent runtime internals.  We now
move all the machine-specific functions to the new header tls_params.h
instead.  The old stubs/tls_*.c files are gone.  For native compilation,
tls_params.h defines trivial inline functions so that they can be optimized
away well in the code using them.

For pnacl, we still need separate, native-compiled functions until the
proper compiler intrinsics get implemented.  Rather than duplicating source
code, these are provided by compiling the tls_params.h inlines as normal
functions (stubs/tls_params.c).  We hope that soon the pnacl compiler
support will be finished properly and we can remove that object file.

BUG=cleanup
TEST=trybots

R=jvoung@google.com

Review URL: http://codereview.chromium.org/7205038

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5721 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2