Fix TLS alignment corner case
In the case where TLS_ALIGNMENT > __nacl_tp_alignment() we need to align
$tp to the larger of the two.  This does not arise in the current code,
but it might soon.

Likewise, when __nacl_tp_alignment() > TLS_ALIGNMENT (as is so on x86),
we want to round up the end of the TLS area to the $tp alignment requirement.

BUG=cleanup
TEST=none

R=jvoung@google.com

Review URL: http://codereview.chromium.org/7217032

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5741 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2