Scons: Move the creation of some trusted envs into functions
This will make it easier to instantiate some of these envs twice, for
x86-32 and x86-64.  Since this differentiation happens early on, in
pre_base_env, I will need to move that instantiation into a function
too, in a later change.

It is also cleaner to move this code out of the Python top level.  It
groups the code and prevents unexpected side-effects via assigning to
global variables.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1988
TEST=trybots

Review URL: http://codereview.chromium.org/7328034

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@5975 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2